+++ .github/workflows/readthedocs.yaml
@@ -17,4 +17,4 @@ jobs:
     steps:
       - uses: readthedocs/actions/preview@v1
         with:
-          project-slug: "bwapy"
+          project-slug: "pybwa"
