+++ README.md
@@ -14,16 +14,16 @@
 [type-checking-link]:   http://mypy-lang.org/
 [pep-8-badge]:          https://img.shields.io/badge/code%20style-pep8-brightgreen.svg
 [pep-8-link]:           https://www.python.org/dev/peps/pep-0008/
-[code-coverage-badge]:  https://codecov.io/gh/fulcrumgenomics/bwapy/branch/main/graph/badge.svg
-[code-coverage-link]:   https://codecov.io/gh/fulcrumgenomics/bwapy
+[code-coverage-badge]:  https://codecov.io/gh/fulcrumgenomics/pybwa/branch/main/graph/badge.svg
+[code-coverage-link]:   https://codecov.io/gh/fulcrumgenomics/pybwa
 [license-badge]:        http://img.shields.io/badge/license-MIT-blue.svg
-[license-link]:         https://github.com/fulcrumgenomics/bwapy/blob/main/LICENSE
+[license-link]:         https://github.com/fulcrumgenomics/pybwa/blob/main/LICENSE
 
-# bwapy
+# pybwa
 
 Python language bindings for [bwa][bwa-link]
 
-See documentation on [bwapy.readthedocs.org][rtd-link].
+See documentation on [pybwa.readthedocs.org][rtd-link].
 
-[rtd-link]: http://bwapy.readthedocs.org/en/stable
+[rtd-link]: http://pybwa.readthedocs.org/en/stable
 [bwa-link]: https://github.com/lh3/bwa].
