--- bwapy/__init__.py
@@ -1,3 +1,0 @@
-from bwapy.libbwaaln import *  # noqa: F403
-from bwapy.libbwaindex import *  # noqa: F403
-from bwapy.libbwamem import *  # noqa: F403
