+++ docs/index.md
@@ -1,9 +1,9 @@
-# bwapy
+# pybwa
 
 Python bindings for BWA
 
 ## Documentation Contents
 
 * [Installation](installation-and-developers-documentation.md)
-* [API](reference/bwapy/index.md)
+* [API](reference/pybwa/index.md)
