+++ pybwa/__init__.py
@@ -1,0 +1,3 @@
+from pybwa.libbwaaln import *  # noqa: F403
+from pybwa.libbwaindex import *  # noqa: F403
+from pybwa.libbwamem import *  # noqa: F403
