+++ pybwa/libbwaaln.pxd
