+++ pybwa/libbwaaln_utils.c
