+++ pybwa/libbwaaln_utils.h
