+++ pybwa/libbwaindex.pxd
