+++ pybwa/libbwaindex.pyi
