+++ pybwa/libbwaindex.pyx
