+++ pybwa/libbwamem.pxd
